misc bug fixes: not allowing heads and corpse to be edited, fixing an error condition in delayed_skin related to corpses, removing excess comments, fixing int/float conversions, and fixing a crashing bug with do_examine that refers to the name of an owned mob.

+++ pp/src/wounds.cpp
@@ -2101,7 +2101,7 @@ do_diagnose (CHAR_DATA * ch, char *argument, int cmd)
 						{
 							sprintf (buf,
 								 "They have lost %d percent of their CON and will regain it in %d in-game hours.\n",
-								 (af->a.spell.sn/ch->con)*100,
+								 (double)(af->a.spell.sn/ch->con)*100,
 								 af->a.spell.duration);
 							send_to_char (buf, ch);
 							continue;
